+++ pyglet/text/layout/scrolling.py
@@ -125,7 +125,7 @@ class ScrollableTextLayout(TextLayout):
 
     def __init__(self, document: AbstractDocument, width: int, height: int, x: float = 0, y: float = 0, z: float = 0,  # noqa: D107
                  anchor_x: AnchorX = "left", anchor_y: AnchorY = "bottom", rotation: float = 0, multiline: bool = False,
-                 dpi: float | None = None, batch: Batch | None = None, group: graphics.Group | None = None,
+                 dpi: int | None = None, batch: Batch | None = None, group: graphics.Group | None = None,
                  program: ShaderProgram | None = None, wrap_lines: bool = True) -> None:
 
         if width is None or height is None:
